* code factoring

+++ scripts/jquery.filetree/connectors/jqueryFileTree.php
@@ -1,6 +1,6 @@
 <?php
 require('../../../connectors/php/filemanager.config.php');
-$root = $_SERVER['DOCUMENT_ROOT'];
+$root = $_SERVER['DOCUMENT_ROOT'].'/';
 
 //
 // jQuery File Tree PHP Connector
@@ -20,24 +20,24 @@ $root = $_SERVER['DOCUMENT_ROOT'];
 //
 //$_POST['dir'] = urldecode($_POST['dir']);
 $_POST['dir'] = str_replace('http://' . $_SERVER['HTTP_HOST'].'/', '', urldecode($_POST['dir']));
-
+$path = str_replace('//','/' , $root . $_POST['dir']); // we remove double slash
 // Check if user is authorized
 if(auth()) {
   
-if( file_exists($root . $_POST['dir']) ) {
-	$files = scandir($root . $_POST['dir']);
+if( file_exists($path) ) {
+	$files = scandir($path);
 	natcasesort($files);
 	if( count($files) > 2 ) { /* The 2 accounts for . and .. */
 		echo "<ul class=\"jqueryFileTree\" style=\"display: none;\">";
 		// All dirs
 		foreach( $files as $file ) {
-			if( file_exists($root . $_POST['dir'] . $file) && !in_array($file, $config['unallowed_dirs']) && $file != '.' && $file != '..' && is_dir($root . $_POST['dir'] . $file) ) {
+			if( file_exists($path . $file) && !in_array($file, $config['unallowed_dirs']) && $file != '.' && $file != '..' && is_dir($root . $_POST['dir'] . $file) ) {
 				echo "<li class=\"directory collapsed\"><a href=\"#\" rel=\"" . htmlentities($_POST['dir'] . $file) . "/\">" . htmlentities($file) . "</a></li>";
 			}
 		}
 		// All files
 		foreach( $files as $file ) {
-			if( file_exists($root . $_POST['dir'] . $file) && !in_array($file, $config['unallowed_files']) && $file != '.' && $file != '..' && !is_dir($root . $_POST['dir'] . $file) ) {
+			if( file_exists($path . $file) && !in_array($file, $config['unallowed_files']) && $file != '.' && $file != '..' && !is_dir($root . $_POST['dir'] . $file) ) {
 				$ext = preg_replace('/^.*\./', '', $file);
 				echo "<li class=\"file ext_$ext\"><a href=\"#\" rel=\"" . htmlentities($_POST['dir'] . $file) . "\">" . htmlentities($file) . "</a></li>";
 			}
